+++ src/components/breadcrumbs/breadcrumbs.style.ts
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+const StyledBreadcrumbs = styled.nav`
+  ol {
+    list-style: none;
+    display: flex;
+    flex-wrap: wrap;
+  }
+`;
+
+export default StyledBreadcrumbs;