+++ src/locales/en-gb.ts
@@ -13,8 +13,16 @@ const enGB: Locale = {
   actionPopover: {
     ariaLabel: () => "actions",
   },
+  advancedColorPicker: {
+    ariaLabel: () => "Change colour",
+    currentColorDescriptionTerm: () => "Current color assigned: ",
+    currentColorAssigned: (currentColor) => currentColor,
+  },
   batchSelection: {
     selected: (count) => `${count} selected`,
+  },
+  breadcrumbs: {
+    ariaLabel: () => "breadcrumbs",
   },
   confirm: {
     no: () => "No",