+++ src/locales/locale.ts
@@ -9,8 +9,16 @@ interface Locale {
   actionPopover: {
     ariaLabel: () => string;
   };
+  advancedColorPicker: {
+    ariaLabel: () => string;
+    currentColorDescriptionTerm: (currentColor: string) => string;
+    currentColorAssigned: (currentColor: string) => string;
+  };
   batchSelection: {
     selected: (count: number | string) => string;
+  };
+  breadcrumbs: {
+    ariaLabel: () => string;
   };
   characterCount: {
     hintString: () => string;